# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Interest Radar API
  version: 1.0.0
  description: |
    사용자의 관심사를 검색 계획으로 변환하고 신규 정보를 수집해 전달하는 API입니다.

    현재는 FID와 Firebase App Check 검증을 임시 비활성화하고 모든 요청을 하나의 익명 설치로
    처리합니다. 명세의 Firebase 보안 header는 기능 복구 시 사용할 계약을 나타냅니다.

    이벤트 polling의 `cursor`는 외부 검색 API의 offset이 아니라 서버 내부의 단조 증가
    이벤트 ID입니다. 응답의 `nextCursor`를 다음 요청에 그대로 사용합니다.
servers:
  - url: http://localhost:3000
    description: Local development
...
paths:
...
  /v1/installations/current:
    put:
      operationId: registerCurrentInstallation
      tags: [Installations]
      summary: 현재 Firebase 설치 등록
      description: FID를 활성화하고 선택적으로 FCM token까지 한 번에 연결합니다.
      security:
        - InstallationIdHeader: []
          FirebaseAppCheck: []
        - {}
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterInstallationRequest"
      responses:
        "200":
          description: 설치 등록 완료
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Installation"
        "400":
          $ref: "#/components/responses/ValidationError"
        "401":
          $ref: "#/components/responses/InstallationAuthenticationError"
    get:
      operationId: getCurrentInstallation
      tags: [Installations]
      summary: 현재 Firebase 설치 조회
      security:
        - InstallationIdHeader: []
          FirebaseAppCheck: []
        - {}
      responses:
        "200":
          description: 활성 설치 정보
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Installation"
        "401":
          $ref: "#/components/responses/InstallationAuthenticationError"
    delete:
      operationId: deleteCurrentInstallation
      tags: [Installations]
      summary: 현재 Firebase 설치와 연결 데이터 해제
      description: 해당 FID의 구독을 중지하고 FCM token을 비활성화합니다.
      security:
        - InstallationIdHeader: []
          FirebaseAppCheck: []
        - {}
      responses:
        "204":
          description: 설치 해제 완료
        "401":
          $ref: "#/components/responses/InstallationAuthenticationError"
  /v1/subscriptions:
    post:
      operationId: createSubscription
      tags: [Subscriptions]
      summary: 관심사 구독 등록
      description: |
        키워드를 AI 또는 deterministic fallback planner로 분석해 검색 계획을 생성합니다.
        검색형 source의 첫 성공 실행은 baseline으로 저장되며 알림 이벤트를 만들지 않습니다.
        이후 신규 검색 후보와 webhook 입력은 AI 검증을 통과해야 polling과 FCM에 노출됩니다.
      security:
        - InstallationIdHeader: []
          FirebaseAppCheck: []
        - {}
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateSubscriptionRequest"
            example:
              keyword: 서울 Java Spring 백엔드 개발자 채용
      responses:
        "201":
          description: 구독 생성 완료
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CreateSubscriptionResponse"
        "400":
          $ref: "#/components/responses/ValidationError"
        "401":
          $ref: "#/components/responses/InstallationAuthenticationError"
        "500":
          $ref: "#/components/responses/InternalError"
    get:
      operationId: listSubscriptions
      tags: [Subscriptions]
      summary: 내 구독 목록 조회
      description: 삭제되지 않은 활성 및 일시정지 구독을 생성 시각 역순으로 반환합니다.
      security:
        - InstallationIdHeader: []
          FirebaseAppCheck: []
        - {}
      responses:
        "200":
          description: 구독 목록
          content:
            application/json:
              schema:
                type: object
                required: [subscriptions]
                properties:
                  subscriptions:
                    type: array
                    items:
                      $ref: "#/components/schemas/Subscription"
        "401":
          $ref: "#/components/responses/InstallationAuthenticationError"
  /v1/subscriptions/{id}:
    parameters:
      - $ref: "#/components/parameters/SubscriptionId"
    get:
      operationId: getSubscription
      tags: [Subscriptions]
      summary: 구독 상세 조회
      security:
        - InstallationIdHeader: []
          FirebaseAppCheck: []
        - {}
      responses:
        "200":
          description: 삭제되지 않은 활성 또는 일시정지 구독 상세
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Subscription"
        "400":
          $ref: "#/components/responses/ValidationError"
        "401":
          $ref: "#/components/responses/InstallationAuthenticationError"
        "404":
          $ref: "#/components/responses/NotFound"
    delete:
      operationId: deleteSubscription
      tags: [Subscriptions]
      summary: 구독 삭제
      description: |
        구독을 soft delete하여 스케줄러와 전송을 중지하고 목록, 상세, 통합 feed에서 제외합니다.
        subscriptions, subscription_events, items 데이터는 DB에 보존됩니다.
      security:
        - InstallationIdHeader: []
          FirebaseAppCheck: []
        - {}
      responses:
        "204":
          description: 구독 soft delete 및 사용자 feed 숨김 완료
        "400":
          $ref: "#/components/responses/ValidationError"
        "401":
          $ref: "#/components/responses/InstallationAuthenticationError"
        "404":
          $ref: "#/components/responses/NotFound"
  /v1/subscriptions/{id}/status:
    parameters:
      - $ref: "#/components/parameters/SubscriptionId"
    patch:
      operationId: updateSubscriptionStatus
      tags: [Subscriptions]
      summary: 구독 일시정지 또는 재개
      description: active=false인 동안 외부 수집, webhook 입력, 신규 feed 생성과 FCM 전송을 중지합니다. 기존 feed는 유지됩니다.
      security:
        - InstallationIdHeader: []
          FirebaseAppCheck: []
        - {}
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateSubscriptionStatusRequest"
            examples:
              pause:
                value:
                  active: false
              resume:
                value:
                  active: true
      responses:
        "204":
          description: 상태 변경 완료
        "400":
          $ref: "#/components/responses/ValidationError"
        "401":
          $ref: "#/components/responses/InstallationAuthenticationError"
        "404":
          $ref: "#/components/responses/NotFound"
  /v1/subscriptions/{id}/events:
    parameters:
      - $ref: "#/components/parameters/SubscriptionId"
      - $ref: "#/components/parameters/Cursor"
      - $ref: "#/components/parameters/Limit"
      - $ref: "#/components/parameters/ProviderFilter"
      - $ref: "#/components/parameters/TextQueryFilter"
      - $ref: "#/components/parameters/FromDateFilter"
      - $ref: "#/components/parameters/ToDateFilter"
      - $ref: "#/components/parameters/BookmarkedFilter"
    get:
      operationId: listSubscriptionEvents
      tags: [Events]
      summary: 특정 구독의 신규 이벤트 polling
      security:
        - InstallationIdHeader: []
          FirebaseAppCheck: []
        - {}
      responses:
        "200":
          description: cursor 이후 이벤트
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EventPage"
        "400":
          $ref: "#/components/responses/ValidationError"
        "401":
          $ref: "#/components/responses/InstallationAuthenticationError"
        "404":
          $ref: "#/components/responses/NotFound"
  /v1/events:
    get:
      operationId: listUserEvents
      tags: [Events]
      summary: 사용자의 전체 신규 이벤트 polling
      description: 삭제되지 않은 모든 구독의 이벤트를 하나의 cursor 흐름으로 반환하는 클라이언트 기본 polling endpoint입니다.
      security:
        - InstallationIdHeader: []
          FirebaseAppCheck: []
        - {}
      parameters:
        - $ref: "#/components/parameters/Cursor"
        - $ref: "#/components/parameters/Limit"
        - $ref: "#/components/parameters/SubscriptionIdFilter"
        - $ref: "#/components/parameters/ProviderFilter"
        - $ref: "#/components/parameters/TextQueryFilter"
        - $ref: "#/components/parameters/FromDateFilter"
        - $ref: "#/components/parameters/ToDateFilter"
        - $ref: "#/components/parameters/BookmarkedFilter"
      responses:
        "200":
          description: cursor 이후 이벤트
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EventPage"
        "400":
          $ref: "#/components/responses/ValidationError"
        "401":
          $ref: "#/components/responses/InstallationAuthenticationError"
  /v1/bookmarks:
    get:
      operationId: listBookmarkedEvents
      tags: [Events]
      summary: 북마크한 이벤트 목록 조회
      description: 삭제되지 않은 구독의 북마크된 이벤트만 하나의 cursor 흐름으로 반환합니다.
      security:
        - InstallationIdHeader: []
          FirebaseAppCheck: []
        - {}
      parameters:
        - $ref: "#/components/parameters/Cursor"
        - $ref: "#/components/parameters/Limit"
        - $ref: "#/components/parameters/SubscriptionIdFilter"
        - $ref: "#/components/parameters/ProviderFilter"
        - $ref: "#/components/parameters/TextQueryFilter"
        - $ref: "#/components/parameters/FromDateFilter"
        - $ref: "#/components/parameters/ToDateFilter"
      responses:
        "200":
          description: 북마크된 이벤트
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EventPage"
        "400":
          $ref: "#/components/responses/ValidationError"
        "401":
          $ref: "#/components/responses/InstallationAuthenticationError"
  /v1/events/{cursor}/bookmark:
    parameters:
      - $ref: "#/components/parameters/EventCursor"
    patch:
      operationId: updateEventBookmark
      tags: [Events]
      summary: 이벤트 북마크 상태 변경
      security:
        - InstallationIdHeader: []
          FirebaseAppCheck: []
        - {}
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateBookmarkRequest"
      responses:
        "204":
          description: 북마크 상태 변경 완료
        "400":
          $ref: "#/components/responses/ValidationError"
        "401":
          $ref: "#/components/responses/InstallationAuthenticationError"
        "404":
          $ref: "#/components/responses/NotFound"
  /v1/devices:
    post:
      operationId: registerDevice
      tags: [Devices]
      summary: FCM 디바이스 토큰 등록
      description: 동일 토큰을 다시 등록하면 사용자, 플랫폼, 활성 상태를 갱신합니다.
      security:
        - InstallationIdHeader: []
          FirebaseAppCheck: []
        - {}
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterDeviceRequest"
      responses:
        "204":
          description: 디바이스 등록 완료
        "400":
          $ref: "#/components/responses/ValidationError"
        "401":
          $ref: "#/components/responses/InstallationAuthenticationError"
    delete:
      operationId: deleteDevice
      tags: [Devices]
      summary: FCM 디바이스 토큰 해제
      security:
        - InstallationIdHeader: []
          FirebaseAppCheck: []
        - {}
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DeleteDeviceRequest"
      responses:
        "204":
          description: 디바이스 비활성화 완료
        "400":
          $ref: "#/components/responses/ValidationError"
        "401":
          $ref: "#/components/responses/InstallationAuthenticationError"
        "404":
          $ref: "#/components/responses/NotFound"
  /v1/webhooks/{subscriptionId}/{source}:
    parameters:
      - $ref: "#/components/parameters/WebhookSubscriptionId"
      - $ref: "#/components/parameters/WebhookSource"
    post:
      operationId: receiveWebhookItems
      tags: [Webhooks]
      summary: 외부 source의 신규 항목 수신
      description: |
        구독 생성 응답으로 받은 secret을 `x-webhook-secret`에 전달합니다.
        같은 source와 item ID가 이미 처리됐다면 `created`에 포함되지 않습니다.
        `publishedAt`이 구독 생성 시각보다 과거인 항목은 저장되지만 사용자 이벤트에서 억제됩니다.
      security:
        - WebhookSecret: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WebhookRequest"
            example:
              items:
                - id: release-42
                  url: https://example.com/releases/42
                  title: Release 42
                  summary: 새로운 버전이 공개되었습니다.
                  publishedAt: "2026-07-10T03:00:00.000Z"
      responses:
        "202":
          description: 요청 처리 완료
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WebhookResponse"
        "400":
          $ref: "#/components/responses/ValidationError"
        "401":
          $ref: "#/components/responses/WebhookAuthenticationError"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"
components:
  securitySchemes:
    InstallationIdHeader:
      type: apiKey
      in: header
      name: x-firebase-installation-id
      description: FID 모드에서 Firebase Installations SDK로 조회한 값입니다. 현재 임시 익명 모드에서는 생략합니다.
    FirebaseAppCheck:
      type: apiKey
      in: header
      name: x-firebase-appcheck
      description: App Check 강제 모드에서 사용하는 token입니다. 현재 임시 익명 모드에서는 생략합니다.
    WebhookSecret:
      type: apiKey
      in: header
      name: x-webhook-secret
      description: 구독 생성 시 한 번 반환되는 webhook secret입니다.
  parameters:
    SubscriptionId:
      name: id
      in: path
      required: true
      description: 구독 UUID
      schema:
        type: string
        format: uuid
    WebhookSubscriptionId:
      name: subscriptionId
      in: path
      required: true
      description: 구독 UUID
      schema:
        type: string
        format: uuid
    WebhookSource:
      name: source
      in: path
      required: true
      description: 이벤트 공급자를 구분하는 이름
      schema:
        type: string
        minLength: 1
        maxLength: 80
      example: default
    EventCursor:
      name: cursor
      in: path
      required: true
      description: 북마크 상태를 변경할 이벤트 cursor
      schema:
        type: integer
        format: int64
        minimum: 1
    Cursor:
      name: cursor
      in: query
      required: false
      description: 마지막으로 처리한 `nextCursor`. 최초 요청은 0입니다.
      schema:
        type: integer
        format: int64
        minimum: 0
        default: 0
    Limit:
      name: limit
      in: query
      required: false
      description: 한 번에 받을 이벤트 수
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 50
    SubscriptionIdFilter:
      name: subscriptionId
      in: query
      required: false
      description: 특정 구독의 이벤트만 조회합니다.
      schema:
        type: string
        format: uuid
    ProviderFilter:
      name: provider
      in: query
      required: false
      description: 저장된 provider 값과 정확히 일치하는 이벤트만 조회합니다.
      schema:
        type: string
        minLength: 1
        maxLength: 120
    TextQueryFilter:
      name: q
      in: query
      required: false
      description: 제목, 요약, URL에 포함된 문자열로 검색합니다.
      schema:
        type: string
        minLength: 1
        maxLength: 300
    FromDateFilter:
      name: from
      in: query
      required: false
      description: publishedAt이 있으면 publishedAt, 없으면 이벤트 생성시각 기준 시작 일시입니다.
      schema:
        type: string
        format: date-time
    ToDateFilter:
      name: to
      in: query
      required: false
      description: publishedAt이 있으면 publishedAt, 없으면 이벤트 생성시각 기준 종료 일시입니다.
      schema:
        type: string
        format: date-time
    BookmarkedFilter:
      name: bookmarked
      in: query
      required: false
      description: 북마크 여부로 필터링합니다.
      schema:
        type: string
        enum: ["true", "false"]
  responses:
    ValidationError:
      description: 경로, query 또는 body 검증 실패
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ValidationError"
    InstallationAuthenticationError:
      description: FID가 없거나 미등록 상태이거나 App Check 검증에 실패함
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/RequestError"
          example:
            error: request_error
            message: x-firebase-installation-id header is required
    WebhookAuthenticationError:
      description: webhook secret이 없거나 일치하지 않음
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/SimpleError"
          example:
            error: invalid_webhook_secret
    NotFound:
      description: 리소스가 없거나 요청 사용자가 소유하지 않음
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/SimpleError"
          example:
            error: not_found
    InternalError:
      description: 서버 내부 오류
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/RequestError"
          example:
            error: internal_error
            message: Unexpected server error
  schemas:
...
    RegisterInstallationRequest:
      type: object
      additionalProperties: false
      required: [platform]
      properties:
        platform:
          type: string
          enum: [ios, android, web]
        fcmToken:
          type: string
          minLength: 20
          maxLength: 4096
    Installation:
      type: object
      additionalProperties: false
      required: [fid, active, createdAt, updatedAt, lastSeenAt]
      properties:
        fid:
          type: string
          minLength: 10
          maxLength: 200
        platform:
          type: string
          enum: [ios, android, web]
        active:
          type: boolean
          const: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        lastSeenAt:
          type: string
          format: date-time
    CreateSubscriptionRequest:
      type: object
      additionalProperties: false
      required: [keyword]
      properties:
        keyword:
          type: string
          minLength: 2
          maxLength: 500
    CreateSubscriptionResponse:
      type: object
      additionalProperties: false
      required: [id, keyword, plan, planner, createdAt, webhook]
      properties:
        id:
          type: string
          format: uuid
        keyword:
          type: string
        plan:
          $ref: "#/components/schemas/SearchPlan"
        planner:
          $ref: "#/components/schemas/PlannerResult"
        createdAt:
          type: string
          format: date-time
        webhook:
          $ref: "#/components/schemas/WebhookRegistration"
    Subscription:
      type: object
      additionalProperties: false
      required: [id, installationId, keyword, plan, active, createdAt, nextRunAt]
      properties:
        id:
          type: string
          format: uuid
        installationId:
          type: string
        keyword:
          type: string
        plan:
          $ref: "#/components/schemas/SearchPlan"
        active:
          type: boolean
        createdAt:
          type: string
          format: date-time
        nextRunAt:
          type: string
          format: date-time
    UpdateSubscriptionStatusRequest:
      type: object
      additionalProperties: false
      required: [active]
      properties:
        active:
          type: boolean
    UpdateBookmarkRequest:
      type: object
      additionalProperties: false
      required: [bookmarked]
      properties:
        bookmarked:
          type: boolean
...
    EventPage:
      type: object
      additionalProperties: false
      required: [events, nextCursor, hasMore]
      properties:
        events:
          type: array
          items:
            $ref: "#/components/schemas/Event"
        nextCursor:
          type: integer
          format: int64
          minimum: 0
          description: 다음 polling 요청에 전달할 cursor
        hasMore:
          type: boolean
          description: 현재 cursor 뒤에 즉시 조회 가능한 이벤트가 더 있는지 여부
...
    RegisterDeviceRequest:
      type: object
      additionalProperties: false
      required: [token, platform]
      properties:
        token:
          type: string
          minLength: 20
          maxLength: 4096
        platform:
          type: string
          enum: [ios, android, web]
    DeleteDeviceRequest:
      type: object
      additionalProperties: false
      required: [token]
      properties:
        token:
          type: string
          minLength: 20
          maxLength: 4096
    WebhookRequest:
      type: object
      additionalProperties: false
      required: [items]
      properties:
        items:
          type: array
          minItems: 1
          maxItems: 100
          items:
            $ref: "#/components/schemas/WebhookItem"
...
    WebhookResponse:
      type: object
      additionalProperties: false
      required: [accepted, created]
      properties:
        accepted:
          type: integer
          minimum: 1
          maximum: 100
          description: 검증을 통과한 입력 항목 수
        created:
          type: integer
          minimum: 0
          maximum: 100
          description: 중복 및 등록 이전 게시물을 제외하고 생성된 사용자 이벤트 수
    ValidationError:
      type: object
      required: [error, details]
      properties:
        error:
          type: string
          const: validation_error
        details:
          type: object
          required: [formErrors, fieldErrors]
          properties:
            formErrors:
              type: array
              items:
                type: string
            fieldErrors:
              type: object
              additionalProperties:
                type: array
                items:
                  type: string
    RequestError:
      type: object
      additionalProperties: false
      required: [error, message]
      properties:
        error:
          type: string
        message:
          type: string
...